Implement screen navigation, display in MenuService
MenuService now displays header before each screen, exit message after last screen.
It displays the latest screen that was pushed to _menuScreens stack.
Navigation controls the stack.
For MenuService to be able to display a specific sceen, its name should be added to MenuScreenName enum and to the switch statement in CreateScreen method. This method creates new instances of screens to be pushed to the stack.