Add SelectTool.cs and SelectOneOptions.cs
SelectTool static class contains SelectOne method that allows user to chose 1 option out of list of strings using Up/Down Arrow to select and Enter/Spacebar to confirm. It returns index of selected option.
This method assumes that each item fits inside 1 row of text. It could be modified to handle longer strings.
SelectOneOptions object passed in optionally allows for customization of selectors - strings displayed at the begining of each row with item and show if the item is selected or not. You can also specify if the selection loops (e.g. if user presses Up Arrow while the first item is selected, the selection moves to the last item).