Fix SaveToJson<T> method in DataAccess
The method threw exception when the save directory didn't exist, so I added check if exists and if it doesn't, the method creates it before trying to write to a file in it.
I also noticed that when running a web app, relative paths point to source code directory instead of the running directory, so I get the actual running directory from AppContext.BaseDirectory.
The paths we had specified with '/'s didn't work with the Path methods, so I had to split the path and recombine it.